Go back to wrap methods, mark wrap/maybe_wrap as pub(crate).
These methods are essentially unsafe when used outside of this
crate as there don't verify the input well enough which can
lead to unsound behavior.

Instead of marking them as unsafe and modifying all of the
callsites, we're just marking them as `pub(crate)`. Maybe
this needs to change in the future. Maybe not?